Split packages and add ObjectLinker test
