+++ programs/README.md
@@ -1,0 +1,3 @@
+# lowarn-programs
+
+See the main README.
