+++ programs/package.yaml
@@ -1,4 +1,4 @@
-name:                lowarn
+name:                lowarn-programs
 version:             0.1.0.0
 github:              "jonathanjameswatson/lowarn"
 license:             BSD3
@@ -8,7 +8,6 @@ copyright:           "2022 Jonathan Watson"
 
 extra-source-files:
 - README.md
-- CHANGELOG.md
 
 # Metadata used when publishing your package
 # synopsis:            Short description of your package
@@ -37,28 +36,8 @@ ghc-options:
 - -Wpartial-fields
 - -Wredundant-constraints
 - -dynamic
+- -shared
+# - -fPIC
 
 library:
   source-dirs: src
-
-executables:
-  lowarn-exe:
-    main:                Main.hs
-    source-dirs:         app
-    ghc-options:
-    - -threaded
-    - -rtsopts
-    - -with-rtsopts=-N
-    dependencies:
-    - lowarn
-
-tests:
-  lowarn-test:
-    main:                Spec.hs
-    source-dirs:         test
-    ghc-options:
-    - -threaded
-    - -rtsopts
-    - -with-rtsopts=-N
-    dependencies:
-    - lowarn
